src/backend/requirements.txt
fastapi
uvicorn
firebase-admin
pymongo
opencv-python
python-multipart
scipy
matplotlib
python-dotenv